refactor: migrate from deprecated VariableStorageUtils to modular system
- Remove deprecated src/lib/core/variable-storage.ts (335 lines)
- Migrate 49 references across 13 files to use modular functions
- Add compatibility bridge for updateVariableUsageCounts()
- Update imports to use specific operation modules
- Maintain full backward compatibility and functionality
- All tests passing, zero breaking changes

Reduces code complexity through 12 focused modules with single responsibilities.
Improves maintainability, type safety, and enables better tree-shaking.
