+++ src/lib/core/index.ts
@@ -3,4 +3,5 @@ export * from './advanced-pattern-matcher';
 export * from './pattern-matcher';
 export * from './storage';
 export * from './variable-resolver';
-export * from './variable-storage';
+// Export the new modular variable storage system
+export * from './variable-storage/index';
